+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/api/ManifestApiCache.kt
@@ -1,4 +1,4 @@
-package com.rakuten.tech.mobile.miniapp.storage
+package com.rakuten.tech.mobile.miniapp.api
 
 import android.content.Context
 import android.content.SharedPreferences
@@ -9,12 +9,10 @@ import com.rakuten.tech.mobile.miniapp.MiniAppManifest
 import kotlin.Exception
 
 /**
- * A caching class to read and store the [MiniAppManifest] per MiniApp's id & version id using [SharedPreferences].
+ * A caching class to read and store the [MiniAppManifest] per miniapp id & version id using [SharedPreferences].
  */
 @Suppress("TooGenericExceptionCaught")
-internal class ManifestApiCache(
-    context: Context
-) {
+internal class ManifestApiCache(context: Context) {
     private val prefs: SharedPreferences = context.getSharedPreferences(
         "com.rakuten.tech.mobile.miniapp.manifest.cache.api", Context.MODE_PRIVATE
     )
@@ -29,11 +27,7 @@ internal class ManifestApiCache(
         }
     }
 
-    fun storeManifest(
-        miniAppId: String,
-        versionId: String,
-        miniAppManifest: MiniAppManifest
-    ) {
+    fun storeManifest(miniAppId: String, versionId: String, miniAppManifest: MiniAppManifest) {
         val jsonToStore: String = Gson().toJson(miniAppManifest)
         prefs.edit().clear().putString(primaryKey(miniAppId, versionId), jsonToStore).apply()
     }